Move _TypeName_from_string
